pyproject.toml
[project]
name = "genai-python-template"
version = "0.1.0"
description = "Production-ready GenAI Python template with Clean Architecture"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    # LLM
    "openai>=1.50.0",
    "instructor>=1.4.0",
    # Vector store
    "chromadb>=0.5.0",
    # Observability
    "structlog>=24.4.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    # Security
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0,<5.0.0",
    # Cache & DB
    "redis>=5.1.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    # Storage
    "aiofiles>=24.1.0",
    # Templating
    "jinja2>=3.1.0",
    # CLI
    "typer>=0.12.0",
    "aiosqlite>=0.20.0",
    # Background Tasks
    "celery[redis]>=5.4.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "pytest-cov>=5.0.0",
    "pip-audit>=2.7.0",
]
ml = [
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
]
data = [
    "pandas>=2.2.0",
    "polars>=1.0.0",
]
spark = [
    "pyspark>=3.5.0",
]
kafka = [
    "aiokafka>=0.11.0",
]
orchestration = [
    "prefect>=3.0.0",
]
quality = [
    "great-expectations>=1.0.0",
    "pandera>=0.20.0",
]
litellm = [
    "litellm>=1.40.0",
]
aws = [
    "aiobotocore>=2.13.0",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/examples/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "C901", # complexity
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "instructor.*",
    "opentelemetry.*",
    "aiofiles.*",
    "celery.*",
    "pandas.*",
    "polars.*",
    "pyspark.*",
    "aiokafka.*",
    "prefect.*",
    "great_expectations.*",
    "pandera.*",
    "litellm.*",
    "aiobotocore.*",
    "botocore.*",
    "jose.*",
    "bcrypt.*",
]
ignore_missing_imports = true

[project.scripts]
genai = "src.interfaces.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]



